:recycle: REV-1821 - updating PR to trigger a single event after collecting data from the form instead of triggering many events whenever a value from the field changes.

+++ data/src/main/java/com/multimoney/data/networking/SmartAccountApi.kt
@@ -57,24 +57,28 @@ class SmartAccountApi @Inject constructor(
         idAddressLevel2: Long,
         idAddressLevel3: Long,
         idEconomicActivity: Long,
-        income: Float,
+        income: Int,
         addressDetail: String,
         isPEP: Boolean,
         user: String,
         idBrand: Int,
         currentStep: String,
     ): ApolloCall<GlobalRequestMutation.Data> =
-        apolloClient.mutation(GlobalRequestMutation(pkUser,
-            status,
-            idProfessionType,
-            idAddressLevel1,
-            idAddressLevel2,
-            idAddressLevel3,
-            idEconomicActivity,
-            income,
-            addressDetail,
-            isPEP,
-            user,
-            idBrand,
-            currentStep)).fetchPolicy(FetchPolicy.NetworkOnly)
+        apolloClient.mutation(
+            GlobalRequestMutation(
+                pkUser,
+                status,
+                idProfessionType,
+                idAddressLevel1,
+                idAddressLevel2,
+                idAddressLevel3,
+                idEconomicActivity,
+                income,
+                addressDetail,
+                isPEP,
+                user,
+                idBrand,
+                currentStep
+            )
+        ).fetchPolicy(FetchPolicy.NetworkOnly)
 }
